Worked on an attempt to implement a ListBox component which nests the TableComponent within it, and adds a query selector to highlight the active row.
In the process, located a bug in the TableComponent. If the keyField is set to null in the headers prop. This is an issue when sending the click item in the table to the handler, as the key field has been excluded from the tableData, therefore cannot be referenced.

This was solved by adding a warning in the useEffect of TableComponent. If the keyfield provided is located in headers and set to null, then throw error that user must provide a different keyField or alternatively be forced to display the keyfrield (make it no null , remove from headers).